// frontend/src/api/generated/ct6.ts
export interface components {
    schemas: {
// ...
        CT6SuccessResponse: {
            api: components["schemas"]["CT6ApiResponseInfo"];
            identity: components["schemas"]["CT6IdentityResponse"];
            learning: components["schemas"]["CT6LearningResponse"];
            post_capacity: components["schemas"]["CT6PostCapacityResponse"];
            pre_capacity: components["schemas"]["CT6PreCapacityResponse"];
            request: components["schemas"]["CT6RequestSummaryResponse"];
            versions: components["schemas"]["CT6VersionsResponse"];
            /** Warnings */
            warnings: components["schemas"]["CT6WarningResponse"][];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}